refactor : 각 Modal 타입 추가
