+++ src/components/Modal/Folder/FolderDeleteModal.tsx
@@ -4,21 +4,25 @@ import {
   ModalSubTitle,
   ModalTitle,
   ModalTitleContainer,
-} from '../../styles/ModalStyle';
-import close_button from '../../assets/svg/close.svg';
+} from '../../../styles/ModalStyle';
+import close_button from '../../../assets/svg/close.svg';
 
-function LinkDeleteModal({ onCloseModal }, link) {
+interface Props {
+  onCloseModal: () => void;
+  name: string;
+}
 
+function FolderDeleteModal({ onCloseModal, name }: Props) {
   return (
     <>
       <CloseButton src={close_button} alt='모달 닫기 버튼' onClick={onCloseModal} />
       <ModalTitleContainer>
-        <ModalTitle>링크 삭제</ModalTitle>
-        <ModalSubTitle>{link}</ModalSubTitle>
+        <ModalTitle>폴더 삭제</ModalTitle>
+        <ModalSubTitle>{name}</ModalSubTitle>
       </ModalTitleContainer>
       <ModalFolderRedButton>삭제하기</ModalFolderRedButton>
     </>
   );
 }
 
-export default LinkDeleteModal;
+export default FolderDeleteModal;
