+++ src/components/Modal/FolderToAddModal.tsx
@@ -1,7 +1,11 @@
 import { CloseButton, InputContainer, ModalFolderButton, ModalInput, ModalTitle } from '../../styles/ModalStyle';
 import close_button from '../../assets/svg/close.svg';
 
-function FolderToAddModal({ onCloseModal }) {
+interface Props {
+  onCloseModal: () => void;
+}
+
+function FolderToAddModal({ onCloseModal }: Props) {
 
   return (
     <>
