+++ src/components/Modal/LinkDeleteModal.tsx
@@ -7,17 +7,23 @@ import {
 } from '../../styles/ModalStyle';
 import close_button from '../../assets/svg/close.svg';
 
-function FolderDeleteModal({ onCloseModal, name }) {
+interface Props {
+  onCloseModal: () => void;
+  link: string;
+}
+
+function LinkDeleteModal({ onCloseModal, link }: Props ) {
+
   return (
     <>
       <CloseButton src={close_button} alt='모달 닫기 버튼' onClick={onCloseModal} />
       <ModalTitleContainer>
-        <ModalTitle>폴더 삭제</ModalTitle>
-        <ModalSubTitle>{name}</ModalSubTitle>
+        <ModalTitle>링크 삭제</ModalTitle>
+        <ModalSubTitle>{link}</ModalSubTitle>
       </ModalTitleContainer>
       <ModalFolderRedButton>삭제하기</ModalFolderRedButton>
     </>
   );
 }
 
-export default FolderDeleteModal;
+export default LinkDeleteModal;
